+++ soffosai/src/core/pipelines/pipeline.mjs
@@ -209,7 +209,7 @@ class Pipeline {
             payload.apiKey = this.apiKey;
 
             let response = await stage.getResponse(payload);
-            if ("error" in response || !isDictObject(response)) {
+            if (response.error || !isDictObject(response)) {
                 infos[stage.name] = response;
                 console.log(response);
                 return infos;
@@ -310,7 +310,7 @@ class Pipeline {
                     let required_key = notation.field;
                     if (reference_service_name == "user_input") {
                         let input_datatype = get_userinput_datatype(user_input[required_key])
-                        if (required_data_type != input_datatype) {
+                        if (required_data_type != input_datatype && required_data_type != 'null') {
                             error_messages.push(`On ${stage.name} service: ${required_data_type} required on user_input '${required_key}' field but ${input_datatype} is provided.`)
                         }
                     } else {
@@ -325,7 +325,7 @@ class Pipeline {
                                 if (output_datatype == 'null') {
                                     error_messages.push(`On ${stage.name} service: the reference service '${reference_service_name}' does not have ${required_key} key on its output.`);
                                 }
-                                if (required_data_type != output_datatype) {
+                                if (required_data_type != output_datatype && required_data_type != 'null') {
                                     error_messages.push(`On ${stage.name} service: The input datatype required for field ${key} is ${required_data_type}. This does not match the datatype to be given by service ${subservice.name}'s ${notation.field} field which is ${output_datatype}.`);
                                 }
                                 break;
@@ -337,7 +337,7 @@ class Pipeline {
                     }
                     
                 } else {
-                    if (get_userinput_datatype(notation) == required_data_type) {
+                    if (get_userinput_datatype(notation) == required_data_type || required_data_type == 'null') {
                         stage._payload[key] = notation;
                     } else {
                         error_messages.push(`On ${stage.name} service: ${key} requires ${required_data_type} but ${typeof notation} is provided.`)