+++ soffosai/src/core/services/documents.mjs
@@ -253,14 +253,14 @@ class DocumentsSearchService extends SoffosAIService {
      * the default operation. If no comparison operator is specified, "$eq" 
      * (or "$in" if the comparison value is a list) is used as the default operation.
      * @param {Array.<string>|InputConfig} [document_ids] - Passing document IDs will confine the search to those documents.
-     * @param {number|InputConfig} [top_n_keywords] - The number of document passages to be retrieved using 
+     * @param {number|InputConfig} [top_n_keywords=5] - The number of document passages to be retrieved using 
      * keyword search. The relevancy is calculated algorithmically based on the frequency of the 
      * query words in the ingested passages. Setting this to 0 disables the keyword search. 
      * When query is left null while top_n_keywords is larger than 0, it will simply filter 
      * the documents based on the rest of the fields like date or metadata. All matched passages will 
      * be returned, therefore the actual value of top_n_keywords does not make a difference, 
      * so long it is larger than 0.
-     * @param {number|InputConfig} [top_n_natural_language] - The number of document passages to be retrieved 
+     * @param {(number|InputConfig)} [top_n_natural_language] - The number of document passages to be retrieved 
      * using Machine Learning-based semantic search. Setting this to 0 disables the semantic search.
      * @param {string|InputConfig} [date_from] - Filters passages to those ingested at or after the specified ISO-8601 formatted date.
      * @param {string|InputConfig} [date_until] - Filters passages to those ingested before the specified ISO-8601 formatted date.